requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn == 0.24.2
joblib
pillow
